+++ tests/testthat/test-coalesce.R
@@ -26,3 +26,28 @@ test_that("coalesce() gives meaningful error messages", {
     coalesce(1:2, letters[1:2])
   })
 })
+
+test_that("coalesce() supports data frames (#5326)", {
+  out <- coalesce(
+    data.frame(x = c(NA, 1)),
+    data.frame(x = 1:2)
+  )
+  expect_identical(out, data.frame(x = c(1, 1)))
+
+  df1 <- data.frame(x = c(NA, 1, NA), y = c(2, NA, NA), z = c(1:2, NA))
+  df2 <- tibble::tibble(x = 1:3, y = c(3, 4, NA), z = c(NA, NA, NA))
+  df3 <- data.frame(x = NA, y = c(30, 40, 50), z = 101:103)
+  out <- coalesce(df1, df2, df3)
+  exp <- tibble(x = c(1, 1, 3), y = c(2, 4, 50), z = c(1L, 2L, 103L))
+  expect_identical(out, exp)
+
+  expect_error(
+    coalesce(
+      data.frame(x = c(NA, 1)),
+      data.frame(x = c("a", "b"))
+    ),
+    class = "vctrs_error_incompatible_type"
+  )
+
+  expect_error(coalesce(as.matrix(mtcars), as.matrix(mtcars)), "matrices")
+})