+++ NEWS.md
@@ -3,6 +3,8 @@
 * `count()` now displays the correct the hint to force a count, `wt = n()` 
   (#5324).
 
+* `coalesce()` now supports data frames correctly (#5326).
+
 * `cummean()` no longer has off-by-one indexing problem (@cropgen, #5287).
 
 # dplyr 1.0.0